+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/ActivitySamplingViewModel.java
@@ -2,6 +2,7 @@ package de.muspellheim.activitysampling.application;
 
 import de.muspellheim.activitysampling.domain.*;
 import de.muspellheim.activitysampling.infrastructure.*;
+import java.time.*;
 import java.time.format.*;
 import java.util.*;
 import javafx.beans.binding.*;
@@ -9,8 +10,16 @@ import javafx.beans.property.*;
 import javafx.collections.*;
 
 class ActivitySamplingViewModel {
+  Runnable onCountdownElapsed;
+
   private final StringProperty activityText = new SimpleStringProperty("");
   private final ReadOnlyBooleanWrapper logButtonDisable = new ReadOnlyBooleanWrapper(true);
+  private final ReadOnlyObjectWrapper<Duration> interval =
+      new ReadOnlyObjectWrapper<>(Duration.ZERO);
+  private final ReadOnlyObjectWrapper<Duration> countdown =
+      new ReadOnlyObjectWrapper<>(Duration.ZERO);
+  private final ReadOnlyStringWrapper countdownLabelText = new ReadOnlyStringWrapper("00:00:00");
+  private final ReadOnlyDoubleWrapper countdownProgress = new ReadOnlyDoubleWrapper(0.0);
   private final ObservableList<ActivityItem> recentActivities = FXCollections.observableArrayList();
 
   private final ActivitiesService activitiesService;
@@ -24,6 +33,22 @@ class ActivitySamplingViewModel {
 
     logButtonDisable.bind(
         Bindings.createBooleanBinding(() -> activityText.get().isBlank(), activityText));
+    countdownLabelText.bind(
+        Bindings.createStringBinding(
+            () -> {
+              var time = LocalTime.ofSecondOfDay(countdown.get().toSeconds());
+              return DateTimeFormatter.ofLocalizedTime(FormatStyle.MEDIUM).format(time);
+            },
+            interval,
+            countdown));
+    countdownProgress.bind(
+        Bindings.createDoubleBinding(
+            () -> {
+              var elapsedSeconds = (double) interval.get().minus(countdown.get()).getSeconds();
+              return elapsedSeconds / interval.get().toSeconds();
+            },
+            interval,
+            countdown));
   }
 
   StringProperty activityTextProperty() {
@@ -34,21 +59,20 @@ class ActivitySamplingViewModel {
     return logButtonDisable.getReadOnlyProperty();
   }
 
+  ReadOnlyStringProperty countdownLabelTextProperty() {
+    return countdownLabelText.getReadOnlyProperty();
+  }
+
+  ReadOnlyDoubleProperty countdownProgressProperty() {
+    return countdownProgress.getReadOnlyProperty();
+  }
+
   ObservableList<ActivityItem> getRecentActivities() {
     return recentActivities;
   }
 
   void run() {
     load();
-  }
-
-  void logActivity() {
-    activitiesService.logActivity(activityText.get());
-    load();
-  }
-
-  void setActivity(Activity activity) {
-    activityText.set(activity.description());
   }
 
   void load() {
@@ -65,5 +89,27 @@ class ActivitySamplingViewModel {
       }
     }
     recentActivities.setAll(items);
+  }
+
+  void logActivity() {
+    activitiesService.logActivity(activityText.get());
+    load();
+  }
+
+  void setActivity(Activity activity) {
+    activityText.set(activity.description());
+  }
+
+  void startCountdown(Duration interval) {
+    this.interval.set(interval);
+    countdown.set(interval);
+  }
+
+  void progressCountdown() {
+    countdown.set(countdown.get().minusSeconds(1));
+    if (countdown.get().isZero()) {
+      Optional.ofNullable(onCountdownElapsed).ifPresent(Runnable::run);
+      countdown.set(interval.get());
+    }
   }
 }